Make `StatusActor` `syncPaths`/`totalPaths` use a set data structure
This avoids recording the same path multiple times. We re-use the same `GitIndexTree` previously used in `Skipper`, renamed to `PathSet`

+++ devbox/common/src/Util.scala
@@ -44,7 +44,7 @@ object Util {
   val timeFormatter = DateTimeFormatter.ofLocalizedDateTime(FormatStyle.SHORT)
     .withZone(ZoneId.systemDefault())
 
-  def formatInt(number: Int) = {
+  def formatInt(number: Long) = {
     val numberFormat = java.text.NumberFormat.getNumberInstance(java.util.Locale.US)
     numberFormat.format(number)
   }
